feat: set header & footer widths for 6.3.8 Activity-Holy Grail
Step 5 - set header and footer widths__

We want to get the header and the footer into the correct position. To
do that, we simply need to make them full width. There are two possible
approaches to accomplishing this:

+ explicitly set one or more of the width properties (`width`,
`min-width`, `max-width`) to the desired value
+ set the `flex-basis`

Either approach will work, though if you go with the first approach you
might also need to set the box-sizing property to be border-box.   We'll
use the flex-basis, since that is simpler and participatory.

Since this is the layout for the entire page, we'll use the vw units to
set the width to be 100 percent of the viewport width.