feat: add code for article & h3 in 4.2.4 Activity
limit the width of each article to 40% of the width of the section, with
justified text. Each should have a margin of 50px all around it, a top
padding of 10px, a left and right padding of 15px and a bottom padding
of 30px. The h3 within each article should be aligned to the right.